Add EnvironmentDefault tests
As a result, we've moved set_env to a generic "helpers" module in our
test code. Further, we also fixed the tests around the upload parsing.
